+++ UIInferno/UICore/_events/WasabyEvents.ts
@@ -12,16 +12,21 @@ import {
     IFixedEvent,
     SwipeController,
     LongTapController,
-    TouchHandlers
+    TouchHandlers, ISyntheticEvent
 } from 'UICommon/Events';
 import {
+    IControlNodeEvent,
     IWasabyHTMLElement,
     TModifyHTMLNode
 } from 'UICommon/interfaces';
 import {
     IDOMEnvironment,
-   IControlNode
+    IControlNode
 } from 'UICore/interfaces';
+import {IArrayEvent} from '../../../UICommon/_events/IEvents';
+
+const callAfterMount: IArrayEvent[] = [];
+const afterMountEvent: string[] = ['mouseenter', 'mousedown'];
 
 export default class WasabyEventsInferno extends WasabyEvents implements IWasabyEventSystem {
     private _environment: IDOMEnvironment;
@@ -57,6 +62,155 @@ export default class WasabyEventsInferno extends WasabyEvents implements IWasaby
         }
     }
 
+    protected vdomEventBubbling<T>(
+        eventObject: ISyntheticEvent,
+        controlNode: T & (IControlNode | IControlNodeEvent),
+        eventPropertiesStartArray: unknown[],
+        args: unknown[],
+        native: boolean
+    ): void {
+        let eventProperties;
+        let stopPropagation = false;
+        const eventPropertyName = 'on:' + eventObject.type.toLowerCase();
+        let curDomNode;
+        let fn;
+        let evArgs;
+        let templateArgs;
+        let finalArgs = [];
+
+        // Если событием стрельнул window или document, то распространение начинаем с body
+        if (native) {
+            curDomNode =
+                eventObject.target === window || eventObject.target === document ? document.body : eventObject.target;
+        } else {
+            curDomNode = controlNode.element;
+        }
+        curDomNode = native ? curDomNode : controlNode.element;
+
+        // Цикл, в котором поднимаемся по DOM-нодам
+        while (!stopPropagation) {
+            eventProperties = curDomNode.eventProperties;
+            if (eventProperties && eventProperties[eventPropertyName]) {
+                // Вызываем обработчики для всех controlNode на этой DOM-ноде
+                const eventProperty = eventPropertiesStartArray || eventProperties[eventPropertyName];
+                for (let i = 0; i < eventProperty.length && !stopPropagation; i++) {
+                    fn = eventProperty[i].fn;
+                    evArgs = eventProperty[i].args || [];
+                    // If controlNode has event properties on it, we have to update args, because of the clos
+                    // happens in template function
+                    templateArgs =
+                        this.isArgsLengthEqual(this.checkControlNodeEvents(controlNode, eventPropertyName, i), evArgs)
+                            ? controlNode.events[eventPropertyName][i].args : evArgs;
+                    try {
+                        if (!args.concat) {
+                            throw new Error(
+                                'Аргументы обработчика события ' + eventPropertyName.slice(3) + ' должны быть массивом.'
+                            );
+                        }
+                        /* Составляем массив аргументов для обаботчика. Первым аргументом будет объект события.
+                         Затем будут аргументы, переданные в обработчик в шаблоне, и последними - аргументы в _notify */
+                        finalArgs = [eventObject];
+                        Array.prototype.push.apply(finalArgs, templateArgs);
+                        Array.prototype.push.apply(finalArgs, args);
+                        // Добавляем в eventObject поле со ссылкой DOM-элемент, чей обработчик вызываем
+                        eventObject.currentTarget = curDomNode;
+
+                        /* Контрол может быть уничтожен, пока его дочернии элементы нотифаят асинхронные события,
+                           в таком случае не реагируем на события */
+                        /* Также игнорируем обработчики контрола, который выпустил событие.
+                         * То есть, сам на себя мы не должны реагировать
+                         * */
+                        if (!fn.control._destroyed && (!controlNode || fn.control !== controlNode.control)) {
+                            try {
+                                // TODO: убрать проверку на тип события - сделать более универсальный метод возможно надо смотреть
+                                //  на eventObject.nativeEvent или вообще для всех?
+                                if (!fn.control._mounted && afterMountEvent.indexOf(eventObject.type) > -1) {
+                                    /* Асинхронный _afterMount контролов приводит к тому,
+                                     * что события с dom начинают стрелять до маунта,
+                                     * в таком случае их надо вызвать отложено */
+                                    callAfterMount.push({fn, finalArgs});
+                                } else {
+                                    let needCallHandler = native;
+                                    if (!needCallHandler) {
+                                        needCallHandler = !this.wasNotified(fn.control._instId, eventObject.type);
+                                        if (needCallHandler && this.needBlockNotifyState() && eventObject.type.indexOf('mouse') === -1) {
+                                            this.setWasNotifyList(fn.control._instId, eventObject.type);
+                                        }
+                                    }
+
+                                    if (needCallHandler) {
+                                        fn.apply(fn.control, finalArgs); // Вызываем функцию из eventProperties
+                                    }
+                                }
+                            } catch (err) {
+                                // в шаблоне могут указать неверное имя обработчика, следует выводить адекватную ошибку
+                                Logger.error(`Ошибка при вызове обработчика "${eventPropertyName}" из контрола ${fn.control._moduleName}.
+                         ${err.message}`, fn.control);
+                            }
+                        }
+                        /* для событий click отменяем стандартное поведение, если контрол уже задестроен.
+                         * актуально для ссылок, когда основное действие делать в mousedown, а он
+                         * срабатывает быстрее click'а. Поэтому контрол может быть уже задестроен
+                         */
+                        if (fn.control._destroyed && eventObject.type === 'click') {
+                            eventObject.preventDefault();
+                        }
+                        /* Проверяем, нужно ли дальше распространять событие по controlNodes */
+                        if (!eventObject.propagating()) {
+                            const needCallNext =
+                                !eventObject.isStopped() &&
+                                eventProperty[i + 1] &&
+                                // при деактивации контролов надо учитывать что событие может распространятся с partial
+                                // если не далать такую проверку то подписка on:deactivated на родителе partial не будет работать
+                                ((eventObject.type === 'deactivated' && eventProperty[i].toPartial) ||
+                                    eventProperty[i + 1].toPartial ||
+                                    eventProperty[i + 1].fn.controlDestination === eventProperty[i].fn.controlDestination);
+                            /* Если подписались на события из HOC'a, и одновременно подписались на контент хока, то прекращать
+                             распространение не нужно.
+                              Пример sync-tests/vdomEvents/hocContent/hocContent */
+                            if (!needCallNext) {
+                                stopPropagation = true;
+                            }
+                        }
+                    } catch (errorInfo) {
+                        let msg = `Event handle: "${eventObject.type}"`;
+                        let errorPoint;
+
+                        if (!fn.control) {
+                            if (typeof window !== 'undefined') {
+                                errorPoint = fn;
+                                msg += '; Error calculating the logical parent for the function';
+                            } else {
+                                errorPoint = curDomNode;
+                            }
+                        } else {
+                            errorPoint = fn.control;
+                        }
+                        Logger.error(msg, errorPoint, errorInfo);
+                    }
+                }
+            }
+            // TODO Remove when compatible is removed
+            if (curDomNode.compatibleNotifier && controlNode && controlNode.element !== curDomNode) {
+                const res = curDomNode.compatibleNotifier.notifyVdomEvent(
+                    eventObject.type,
+                    args,
+                    controlNode && controlNode.control
+                );
+                if (!eventObject.hasOwnProperty('result')) {
+                    eventObject.result = res;
+                }
+            }
+            curDomNode = curDomNode.parentNode;
+            if (curDomNode === null || curDomNode === undefined || !eventObject.propagating()) {
+                stopPropagation = true;
+            }
+            if (eventPropertiesStartArray !== undefined) {
+                eventPropertiesStartArray = undefined;
+            }
+        }
+    }
+
     private needPropagateEvent(environment: IDOMEnvironment, event: IFixedEvent): boolean {
         if (!environment._rootDOMNode) {
             return false;
@@ -80,9 +234,9 @@ export default class WasabyEventsInferno extends WasabyEvents implements IWasaby
             // И чтобы обработчики событий не были вызваны два раза, стопаем нативное событие.
             return false;
         } else if (detection.isMobileIOS &&
-           FastTouchEndController.isFastEventFired(event.type) &&
-           FastTouchEndController.wasEventEmulated() &&
-           event.isTrusted) {
+            FastTouchEndController.isFastEventFired(event.type) &&
+            FastTouchEndController.wasEventEmulated() &&
+            event.isTrusted) {
             // на ios 14.4 после событий тача стреляет дополнительный mousedown с isTrusted = true
             // это связанно с тем, что мы пытаемся игнорировать нативную задержку в 300 мс
             // поэтому для событий которые мы выстрелим руками повторный вызов не нужен
@@ -93,6 +247,23 @@ export default class WasabyEventsInferno extends WasabyEvents implements IWasaby
 
         return true;
     }
+
+    callEventsToDOM(): void {
+        while (callAfterMount && callAfterMount.length) {
+            const elem = callAfterMount.shift();
+            const fn = elem.fn;
+            /* в слое совместимости контрол внутри которого построился wasaby-контрол, может быть уничтожен
+               до того как начнется асинхронный вызов afterMount,
+               как результат в текущей точку контрол будет уже уничтожен слоем совместимости
+               нало проверять действительно ли он жив, перед тем как выстрелить событием
+               */
+            // @ts-ignore
+            if (!fn.control._destroyed) {
+                fn.apply(fn.control, elem.finalArgs);
+            }
+        }
+    }
+
     //#endregion
 
     //#region события тача
@@ -115,6 +286,7 @@ export default class WasabyEventsInferno extends WasabyEvents implements IWasaby
         };
         LongTapController.initState(event, longTapCallback.bind(this._environment));
     }
+
     //#endregion
 
     //#region _notify события
@@ -145,6 +317,7 @@ export default class WasabyEventsInferno extends WasabyEvents implements IWasaby
         this.clearWasNotifyList();
         return eventObject.result;
     }
+
     //#endregion
 
     //#region хэлперы
@@ -190,6 +363,7 @@ export default class WasabyEventsInferno extends WasabyEvents implements IWasaby
         }
         return result;
     }
+
     //#endregion
 
     //#region регистрация событий
@@ -216,6 +390,7 @@ export default class WasabyEventsInferno extends WasabyEvents implements IWasaby
             this.addHandler(eventName, false, handler, true);
         }
     }
+
     //#endregion
 }
