+++ UIReact/UICore/_events/WasabyEvents.ts
@@ -11,7 +11,7 @@ import {
     LongTapController,
     IWasabyEvent,
     EventUtils,
-    TouchHandlers
+    TouchHandlers, ISyntheticEvent
 } from 'UICommon/Events';
 import {
     IWasabyHTMLElement,
@@ -21,6 +21,7 @@ import {
 import { WasabyEventsDebug } from './WasabyEventsDebug';
 import { Control } from 'UICore/Control';
 import { Set } from 'Types/shim';
+import {Logger} from '../../../UICommon/Utils';
 
 type TElement =  HTMLElement & {
     eventProperties?: Record<string, IWasabyEvent[]>;
@@ -67,6 +68,147 @@ export default class WasabyEventsReact extends WasabyEvents implements IWasabyEv
         }
         this.vdomEventBubbling(syntheticEvent, null, undefined, [], true);
     }
+
+    protected vdomEventBubbling<T>(
+        eventObject: ISyntheticEvent,
+        controlNode: T & IControlNodeEvent,
+        eventPropertiesStartArray: unknown[],
+        args: unknown[],
+        native: boolean
+    ): void {
+        let eventProperties;
+        let stopPropagation = false;
+        const eventPropertyName = 'on:' + eventObject.type.toLowerCase();
+        let curDomNode;
+        let fn;
+        let evArgs;
+        let templateArgs;
+        let finalArgs = [];
+
+        // Если событием стрельнул window или document, то распространение начинаем с body
+        if (native) {
+            curDomNode =
+                eventObject.target === window || eventObject.target === document ? document.body : eventObject.target;
+        } else {
+            curDomNode = controlNode.element;
+        }
+        curDomNode = native ? curDomNode : controlNode.element;
+
+        // Цикл, в котором поднимаемся по DOM-нодам
+        while (!stopPropagation) {
+            eventProperties = curDomNode.eventProperties;
+            if (eventProperties && eventProperties[eventPropertyName]) {
+                // Вызываем обработчики для всех controlNode на этой DOM-ноде
+                const eventProperty = eventPropertiesStartArray || eventProperties[eventPropertyName];
+                for (let i = 0; i < eventProperty.length && !stopPropagation; i++) {
+                    fn = eventProperty[i].fn;
+                    evArgs = eventProperty[i].args || [];
+                    // If controlNode has event properties on it, we have to update args, because of the clos
+                    // happens in template function
+                    templateArgs =
+                        this.isArgsLengthEqual(this.checkControlNodeEvents(controlNode, eventPropertyName, i), evArgs)
+                            ? controlNode.events[eventPropertyName][i].args : evArgs;
+                    try {
+                        if (!args.concat) {
+                            throw new Error(
+                                'Аргументы обработчика события ' + eventPropertyName.slice(3) + ' должны быть массивом.'
+                            );
+                        }
+                        /* Составляем массив аргументов для обаботчика. Первым аргументом будет объект события.
+                         Затем будут аргументы, переданные в обработчик в шаблоне, и последними - аргументы в _notify */
+                        finalArgs = [eventObject];
+                        Array.prototype.push.apply(finalArgs, templateArgs);
+                        Array.prototype.push.apply(finalArgs, args);
+                        // Добавляем в eventObject поле со ссылкой DOM-элемент, чей обработчик вызываем
+                        eventObject.currentTarget = curDomNode;
+
+                        /* Контрол может быть уничтожен, пока его дочернии элементы нотифаят асинхронные события,
+                           в таком случае не реагируем на события */
+                        /* Также игнорируем обработчики контрола, который выпустил событие.
+                         * То есть, сам на себя мы не должны реагировать
+                         * */
+                        if (!fn.control._destroyed && (!controlNode || fn.control !== controlNode.control)) {
+                            try {
+                                let needCallHandler = native;
+                                if (!needCallHandler) {
+                                    needCallHandler = !this.wasNotified(fn.control._instId, eventObject.type);
+                                    if (needCallHandler && this.needBlockNotifyState() && eventObject.type.indexOf('mouse') === -1) {
+                                        this.setWasNotifyList(fn.control._instId, eventObject.type);
+                                    }
+                                }
+
+                                if (needCallHandler) {
+                                    fn.apply(fn.control, finalArgs); // Вызываем функцию из eventProperties
+                                }
+                            } catch (err) {
+                                // в шаблоне могут указать неверное имя обработчика, следует выводить адекватную ошибку
+                                Logger.error(`Ошибка при вызове обработчика "${eventPropertyName}" из контрола ${fn.control._moduleName}.
+                         ${err.message}`, fn.control);
+                            }
+                        }
+                        /* для событий click отменяем стандартное поведение, если контрол уже задестроен.
+                         * актуально для ссылок, когда основное действие делать в mousedown, а он
+                         * срабатывает быстрее click'а. Поэтому контрол может быть уже задестроен
+                         */
+                        if (fn.control._destroyed && eventObject.type === 'click') {
+                            eventObject.preventDefault();
+                        }
+                        /* Проверяем, нужно ли дальше распространять событие по controlNodes */
+                        if (!eventObject.propagating()) {
+                            const needCallNext =
+                                !eventObject.isStopped() &&
+                                eventProperty[i + 1] &&
+                                // при деактивации контролов надо учитывать что событие может распространятся с partial
+                                // если не далать такую проверку то подписка on:deactivated на родителе partial не будет работать
+                                ((eventObject.type === 'deactivated' && eventProperty[i].toPartial) ||
+                                    eventProperty[i + 1].toPartial ||
+                                    eventProperty[i + 1].fn.controlDestination === eventProperty[i].fn.controlDestination);
+                            /* Если подписались на события из HOC'a, и одновременно подписались на контент хока, то прекращать
+                             распространение не нужно.
+                              Пример sync-tests/vdomEvents/hocContent/hocContent */
+                            if (!needCallNext) {
+                                stopPropagation = true;
+                            }
+                        }
+                    } catch (errorInfo) {
+                        let msg = `Event handle: "${eventObject.type}"`;
+                        let errorPoint;
+
+                        if (!fn.control) {
+                            if (typeof window !== 'undefined') {
+                                errorPoint = fn;
+                                msg += '; Error calculating the logical parent for the function';
+                            } else {
+                                errorPoint = curDomNode;
+                            }
+                        } else {
+                            errorPoint = fn.control;
+                        }
+                        Logger.error(msg, errorPoint, errorInfo);
+                    }
+                }
+            }
+            // TODO Remove when compatible is removed
+            if (curDomNode.compatibleNotifier && controlNode && controlNode.element !== curDomNode) {
+                const res = curDomNode.compatibleNotifier.notifyVdomEvent(
+                    eventObject.type,
+                    args,
+                    controlNode && controlNode.control
+                );
+                if (!eventObject.hasOwnProperty('result')) {
+                    eventObject.result = res;
+                }
+            }
+            curDomNode = curDomNode.parentNode;
+            if (curDomNode === null || curDomNode === undefined || !eventObject.propagating()) {
+                stopPropagation = true;
+            }
+            if (eventPropertiesStartArray !== undefined) {
+                eventPropertiesStartArray = undefined;
+            }
+        }
+    }
+
     //#endregion
 
     //#region события тача
